refactor: extract urlSchema and paramsSchema to useURLAndParamsSchema
